+++ http/src/main/scala/com/twitter/finatra/http/filters/TraceIdMDCFilter.scala
@@ -7,6 +7,9 @@ import javax.inject.Singleton
 import org.slf4j.MDC
 
 /**
+ * Filter to log tracing data into MDC bag of attributes.
+ * Includes traceSampled flag that indicates if the trace
+ * is available via tracers
  * Note: Any MDC filter must be used in conjunction with the LoggingMDCFilter
  * to ensure that diagnostic context is properly managed.
  */
@@ -14,7 +17,11 @@ import org.slf4j.MDC
 class TraceIdMDCFilter[Req, Rep] extends SimpleFilter[Req, Rep] {
 
   override def apply(request: Req, service: Service[Req, Rep]): Future[Rep] = {
-    MDC.put("traceId", Trace.id.traceId.toString())
+    MDC.put("traceId", Trace.id.traceId.toString)
+    // If sampling decision is not made yet
+    // consider the trace not sampled for this span scope
+    MDC.put("traceSampled", Trace.id._sampled.getOrElse(false).toString)
+    MDC.put("traceSpanId", Trace.id.spanId.toString)
     service(request)
   }
 }